api/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["CoSm1C AvEnGeR <zhora.zhilin.06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.1"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
httpx-oauth = "^0.14.0"
starlette-exporter = "^0.21.0"
python-dotenv = "^1.0.1"
strawberry-graphql = {extras = ["fastapi"], version = "^0.247.0"}
pendulum = "^3.0.0"
gunicorn = "^21.2.0"
redis = "^5.0.3"
python-multipart = "^0.0.10"
groq = "^0.8.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
ffmpeg-python = "^0.2.0"
pydub = "^0.25.1"
aiodataloader = "^0.4.0"
python-logstash = "^0.4.8"
aiosmtplib = "^3.0.1"
faststream = {extras = ["asyncio"], version = "^0.5.15"}
nats-py = "^2.8.0"
google-api-python-client = "^2.139.0"
google-auth-oauthlib = "^1.2.1"
aiohttp = "^3.10.3"
bcrypt = "^4.2.0"
pytz = "^2024.1"
dishka = "^1.3.0"
argon2-cffi = "^23.1.0"  # Оставляем в основных зависимостях
boto3 = "^1.35.24"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-bugbear = "^24.4.26"
flake8-comprehensions = "^3.15.0"
flake8-quotes = "^3.4.0"
isort = "^5.13.2"
mypy = "^1.11.1"
autoflake = "^2.3.1"
ruff = "^0.6.0"
black = "^24.8.0"
pre-commit = "^3.8.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
psycopg2-binary = "^2.9.8" # Добавляем psycopg2 для синхронного подключения в тестах
requests-toolbelt = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']
exclude = '''
/(
    api/migrations
    | tg-bot
)/
'''

[tool.ruff]
line-length = 89
target-version="py311"
src = ["src"]
exclude = [
    "migrations/",
]
include = ["src/**.py", "tests/**.py"]

[tool.ruff.lint]
ignore = [
   "ARG",
   "ANN",
   "D",
   "EM101",
   "EM102",
   "PT001",
   "PT023",
   "SIM108",
   "SIM114",
   "TRY003",
   "PLW2901",
   "RET505",
   "PLR0913",
   "UP038",
   "TCH001",
   "SLF001",
   "COM812",
   "ISC001",
   "S101",
   "RUF002",
   "B008",
   "E721",
   "INP001",
   "ERA001", #  проверка на докстринги
   "RUF001", #  кириллица в докстрингах
   "RUF003" #  кириллица в комментах
]
select = [
  "ALL",
]

[tool.ruff.lint.isort]
no-lines-before = ["local-folder"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = "ignore::DeprecationWarning"
